backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Kanak Tanwar",email = "kanaktanwarpro@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "pydantic-settings (>=2.8.0,<3.0.0)",
    "pymongo (>=4.11.1,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
]


[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
